+++ R/summarise_tb_burden.R
@@ -26,7 +26,7 @@
 #' @export
 #'
 #' @import magrittr
-#' @importFrom dplyr mutate group_by ungroup select select_at mutate_at funs left_join lag bind_rows summarise summarise_at one_of rename_at arrange n contains
+#' @importFrom dplyr mutate group_by ungroup select select_at mutate_at left_join lag bind_rows summarise summarise_at one_of rename_at arrange n contains
 #' @importFrom purrr map map2_dfr compact reduce map_lgl
 #' @importFrom tibble as_tibble
 #' @importFrom tidyr nest unnest
@@ -94,9 +94,8 @@ summarise_tb_burden <- function(df = NULL,
                                 annual_change = FALSE,
                                 download_data = TRUE,
                                 save = TRUE,
-                                burden_save_name = "TB_burden",
-                                dict_save_name = "TB_data_dict",
-                                verbose = TRUE) {
+                                verbose = TRUE,
+                                ...) {
   
   ## Deal with undefined global function notes
   . <- NULL; Area <- NULL; Year <- NULL; country <- NULL; data <- NULL; e_pop_num <- NULL;
@@ -126,9 +125,9 @@ summarise_tb_burden <- function(df = NULL,
       } else if (stat %in% c("rate", "prop")) {
         summarised_df <- summarised_df %>% 
           summarise_at(.vars = c(metrics, "denom"),
-                       funs(sum(as.numeric(.), na.rm = T))) %>% 
+                       list(~ sum(as.numeric(.), na.rm = T))) %>% 
           mutate_at(.vars = int_metrics,
-                    .funs = funs(. / denom * int_rate_scale)) %>% 
+                    .funs = list(~ . / denom * int_rate_scale)) %>% 
           select(-denom)
         
         
@@ -156,9 +155,8 @@ summarise_tb_burden <- function(df = NULL,
                                     annual_change = FALSE,
                                     download_data = download_data,
                                     save = save,
-                                    burden_save_name = burden_save_name,
-                                    dict_save_name = dict_save_name,
-                                    verbose = verbose )$df
+                                    verbose = verbose,
+                                    ...)$df
     
     countries_df <- mutate(countries_df, Area = as.character(country))
   }else{
@@ -183,9 +181,8 @@ summarise_tb_burden <- function(df = NULL,
                                   annual_change = FALSE,
                                   download_data = download_data,
                                   save = save,
-                                  burden_save_name = burden_save_name,
-                                  dict_save_name = dict_save_name,
-                                  verbose = verbose )$df
+                                  verbose = verbose,
+                                  ...)$df
     
     regions_df <- mutate(regions_df, Area = as.character(g_whoregion))
   }else{
@@ -204,9 +201,8 @@ summarise_tb_burden <- function(df = NULL,
                                 annual_change = FALSE,
                                 download_data = download_data,
                                 save = save,
-                                burden_save_name = burden_save_name,
-                                dict_save_name = dict_save_name,
-                                verbose = verbose )$df
+                                verbose = verbose,
+                                ...)$df
     
     world_df <- mutate(world_df, Area = "Global")
     
@@ -238,9 +234,8 @@ summarise_tb_burden <- function(df = NULL,
                                                         annual_change = FALSE,
                                                         download_data = download_data,
                                                         save = save,
-                                                        burden_save_name = burden_save_name,
-                                                        dict_save_name = dict_save_name,
-                                                        verbose = verbose)$df %>% 
+                                                        verbose = verbose,
+                                                        ...)$df %>% 
                  mutate(Area = .y)
       )
     )
@@ -294,7 +289,7 @@ summarise_tb_burden <- function(df = NULL,
     } else if (is.null(conf)) {
 
       summarised_df <-  all_df %>% 
-        rename_at(.vars = metric, .funs = funs(paste0("samples")))
+        rename_at(.vars = metric, .funs = list( ~ paste0("samples")))
   
       
     }else{
@@ -304,7 +299,7 @@ summarise_tb_burden <- function(df = NULL,
       ## If the data comes with confidence intervals attached
       summarised_df <- all_df %>% 
         mutate_at(.vars = metrics, 
-                  .funs = funs(ifelse(is.na(.), all_df[[metric]], .))) %>% 
+                  .funs = list( ~ ifelse(is.na(.), all_df[[metric]], .))) %>% 
         group_by(Area, Year)
       
       summarised_df$sd <- (summarised_df[[metrics[3]]] - summarised_df[[metrics[2]]]) / (2 * 1.96)
@@ -333,7 +328,7 @@ summarise_tb_burden <- function(df = NULL,
       get_summary() %>% 
       ungroup %>% 
       mutate_at(.vars = c("mean", "mean_lo", "mean_hi"),
-                .funs = funs(ifelse(. %in% NaN, NA, .)))
+                .funs = list(~ ifelse(. %in% NaN, NA, .)))
 
     
     ## Clean up summarised results
@@ -343,7 +338,7 @@ summarise_tb_burden <- function(df = NULL,
     if (truncate_at_zero) {
       summarised_df <- summarised_df %>% 
         mutate_at(.vars = c("mean", "mean_lo", "mean_hi"),
-                  .funs = funs(ifelse(. < 0, 0, .)))
+                  .funs = list(~ ifelse(. < 0, 0, .)))
     }
     
     colnames(summarised_df) <- c("area", "year", paste0(metric, c("", "_lo", "_hi")))
@@ -377,7 +372,7 @@ summarise_tb_burden <- function(df = NULL,
        get_summary() %>% 
        ungroup %>% 
        mutate_at(.vars = c("mean", "mean_lo", "mean_hi"),
-                 .funs = funs(ifelse(. %in% NaN, NA, .)))
+                 .funs = list(~ ifelse(. %in% NaN, NA, .)))
      
      colnames(output_df) <- c("area", "year", paste0(metric, c("", "_lo", "_hi")))
    }
@@ -408,12 +403,12 @@ summarise_tb_burden <- function(df = NULL,
     output_df <- output_df %>% 
       group_by(area) %>% 
       arrange(year) %>% 
-      mutate_at(.vars = metrics, .funs = funs((. - lag(.)) / lag(.))) %>%
+      mutate_at(.vars = metrics, .funs = list( ~ (. - lag(.)) / lag(.))) %>%
       arrange(year) %>% 
       slice(-1) %>% 
       ungroup %>% 
       mutate_at(.vars = metrics,
-                .funs = funs(replace(., is.nan(.), NA)))
+                .funs = list( ~ replace(., is.nan(.), NA)))
   }
   
   return(output_df)