+++ tests/testthat/test-get_tb_burden.R
@@ -53,3 +53,32 @@ test_that("When TB burden data can be correctly downloaded when MDR is not also 
 
   expect_true(ncol(tb_burden_with_mdr) > ncol(tb_burden_no_mdr))
 })
+
+
+test_that("Additional datasets can all be downloaded", {
+  skip_on_cran()
+  expect_true(ncol(tb_data) < ncol(get_tb_burden(additional_datasets = "all")))
+})
+
+test_that("Additional datasets have no columns that are  misjoined", {
+  skip_on_cran()
+  expect_equal(ncol(dplyr::select(get_tb_burden(additional_datasets = "all"), 
+                                  dplyr::ends_with(".x"), 
+                                  dplyr::ends_with(".y"))), 
+               c(as.integer(0)))
+})
+
+test_that("A single additional dataset can be downloaded", {
+  skip_on_cran()
+  expect_true(ncol(tb_data) < 
+                ncol(get_tb_burden(
+    additional_datasets = available_datasets$dataset[3])))
+})
+
+
+test_that("Downloading fails when an unsupported dataset name is given", {
+  skip_on_cran()
+  expect_error(get_tb_burden(additional_datasets = "gwegwefgw"))
+})
+
+