+++ tests/testthat/test-plot_tb_burden.R
@@ -10,21 +10,21 @@ test_that("plot_tb_burden produces a plot", {
 })
 
 test_that("plot_tb_burden produces a plot with smoothed data", {
-  plot <- plot_tb_burden(download_data = TRUE, save = TRUE, smooth = TRUE)
+  plot <- suppressWarnings(plot_tb_burden(download_data = TRUE, save = TRUE, smooth = TRUE))
   
-  expect_true(!is.null(plot))
-  expect_equal("ggplot", class(plot)[2])
+  expect_true(!is.null(suppressMessages(suppressWarnings(plot))))
+  expect_equal("ggplot", class(suppressMessages(suppressWarnings(plot)))[2])
   skip_on_cran()
-  vdiffr::expect_doppelganger("smooth-plot", plot)
+  vdiffr::expect_doppelganger("smooth-plot", suppressMessages(suppressWarnings(plot)))
 })
 
 test_that("plot_tb_burden produces a plot with a log10 transform", {
   plot <- plot_tb_burden(trans = "log10")
   
-  expect_true(!is.null(plot))
-  expect_equal("ggplot", class(plot)[2])
+  expect_true(!is.null(suppressMessages(suppressWarnings(plot))))
+  expect_equal("ggplot", class(suppressMessages(suppressWarnings(plot)))[2])
   skip_on_cran()
-  vdiffr::expect_doppelganger("log-plot", plot)
+  vdiffr::expect_doppelganger("log-plot", suppressMessages(suppressWarnings(plot)))
 })
 
 test_that("plot_tb_burden produces a plot with the annual_change option", {